+++ packages/components/src/interface/index.ts
@@ -9,7 +9,6 @@
  * Copyright (C) The OpenCRVS Authors. OpenCRVS and the OpenCRVS
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
-export * from './Box'
 export * from './Header'
 export * from './Tabs'
 export * from './ActionPageLight'
